requirements.txt
requests~=2.27.1
python-dotenv~=0.20.0
python-telegram-bot-calendar~=1.0.5
pyTelegramBotAPI~=4.5.0

